requirements.txt
anthropic
requests
python-dotenv
feedparser
python-telegram-bot
